+++ src/document/query/find/ancestor/find-ancestor.types.ts
@@ -3,4 +3,5 @@ import type { AnyDialecteConfig, ElementsOf } from '@/types'
 export type FindAncestorsOptions<GenericConfig extends AnyDialecteConfig> = {
 	depth?: number
 	stopAtTagName?: ElementsOf<GenericConfig>
+	order?: 'bottom-up' | 'top-down'
 }